Change Integer to Long in Product.java

+++ src/main/java/dev/nmarulo/depensaapp/app/shoppinglist/dtos/DeleteProductsShoppingListReq.java
@@ -7,6 +7,6 @@ import java.util.List;
 @Data
 public class DeleteProductsShoppingListReq {
     
-    private List<Integer> productsId;
+    private List<Long> productsId;
     
 }
